+++ src/components/Filters/Filters.tsx
@@ -6,20 +6,12 @@ import { textSans } from "@guardian/src-foundations/typography";
 
 import { Pagination } from "../Pagination/Pagination";
 
+import { FilterOptions, OrderByType, ThreadsType } from "../../types";
+
 type Props = {
   filters: FilterOptions;
   setFilters: (newFilterObject: FilterOptions) => void;
   pages: number;
-};
-
-type orderByType = "newest" | "oldest" | "mostrecommended";
-type threadsType = "collapsed" | "expanded" | "unthreaded";
-
-export type FilterOptions = {
-  orderBy: orderByType;
-  pageSize: number;
-  threads: threadsType;
-  page: number;
 };
 
 export const defaultFilterOptions: FilterOptions = {
@@ -87,7 +79,7 @@ export const Filters = ({ filters, setFilters, pages }: Props) => {
             onChange={e =>
               setFilters({
                 ...filters,
-                orderBy: e.target.value as orderByType
+                orderBy: e.target.value as OrderByType
               })
             }
             value={filters.orderBy}
@@ -127,7 +119,7 @@ export const Filters = ({ filters, setFilters, pages }: Props) => {
             onChange={e =>
               setFilters({
                 ...filters,
-                threads: e.target.value as threadsType
+                threads: e.target.value as ThreadsType
               })
             }
             value={filters.threads}
